Fixes performance issue https://github.com/denoland/deno/issues/12090

Added benchmark `response_string` which produced the following results before and after the change:

* using USVString:
```
Benchmark #13: /home/luismalheiro/projects/deno_pr/target/release/deno run cli/tests/testdata/response_string_perf.js 
  Time (mean ± σ):     432.9 ms ±  26.8 ms    [User: 308.8 ms, System: 192.8 ms]
  Range (min … max):   409.5 ms … 485.4 ms    10 runs
```
* using DOMString:
```
Benchmark #13: /home/luismalheiro/projects/deno_pr/target/release/deno run cli/tests/testdata/response_string_perf.js 
  Time (mean ± σ):     175.8 ms ±  17.6 ms    [User: 168.6 ms, System: 73.9 ms]
  Range (min … max):   147.3 ms … 211.6 ms    18 runs
```

Notice that the difference will increase non-linearly for large strings.